requirements.txt
certifi==2021.10.8
click==8.0.3
colorama==0.4.4
Flask==2.0.2
grpcio==1.41.1
grpcio-tools==1.41.1
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
protobuf==3.19.1
six==1.16.0
Werkzeug==2.0.2
wincertstore==0.2
